Refactor: Split client service into modules
Restructure `src/services/client-service.ts` into smaller, modular files (e.g., profile-service.ts, workout-service.ts, auth-service.ts) to improve code organization and maintainability. Functionality remains unchanged.
