+++ src/services/clients/auth.ts
@@ -1,8 +1,28 @@
 
-import { supabase } from '@/integrations/supabase/client';
+import { supabase } from "@/integrations/supabase/client";
 
 /**
- * Send a password reset email to a user
+ * Delete a user account (admin only)
+ */
+export const deleteUser = async (userId: string): Promise<boolean> => {
+  try {
+    const { data, error } = await supabase
+      .rpc('admin_delete_user', { user_id: userId });
+
+    if (error) {
+      console.error('Error deleting user:', error);
+      return false;
+    }
+
+    return true;
+  } catch (error) {
+    console.error('Error deleting user:', error);
+    return false;
+  }
+};
+
+/**
+ * Send password reset email
  */
 export const sendPasswordResetEmail = async (email: string): Promise<boolean> => {
   try {
@@ -14,28 +34,10 @@ export const sendPasswordResetEmail = async (email: string): Promise<boolean> =>
       console.error('Error sending password reset email:', error);
       return false;
     }
-    
-    return true;
-  } catch (error) {
-    console.error('Error in sendPasswordResetEmail:', error);
-    return false;
-  }
-};
 
-/**
- * Delete a user (Admin only)
- */
-export const deleteUser = async (userId: string): Promise<boolean> => {
-  try {
-    // Admin function to delete a user
-    // This would be implemented as needed
-    
-    console.log('Deleting user:', userId);
-    
-    // Return implementation would go here
     return true;
   } catch (error) {
-    console.error('Error deleting user:', error);
+    console.error('Error sending password reset email:', error);
     return false;
   }
 };
