+++ src/services/clients/workout.ts
@@ -1,8 +1,52 @@
 
-import { supabase } from '@/integrations/supabase/client';
+import { supabase } from "@/integrations/supabase/client";
 
 /**
- * Track a workout set for a client
+ * Saves workout journal notes for a specific workout completion
+ */
+export const saveWorkoutJournalNotes = async (workoutCompletionId: string, notes: string): Promise<boolean> => {
+  try {
+    const { error } = await supabase
+      .from('workout_completions')
+      .update({ notes: notes })
+      .eq('id', workoutCompletionId);
+
+    if (error) {
+      console.error('Error saving workout journal notes:', error);
+      return false;
+    }
+
+    return true;
+  } catch (error) {
+    console.error('Error saving workout journal notes:', error);
+    return false;
+  }
+};
+
+/**
+ * Updates the completion date of a workout
+ */
+export const updateWorkoutCompletionDate = async (workoutId: string, newDate: Date): Promise<boolean> => {
+  try {
+    const { error } = await supabase
+      .from('workout_completions')
+      .update({ completed_at: newDate.toISOString() })
+      .eq('id', workoutId);
+
+    if (error) {
+      console.error('Error updating workout date:', error);
+      return false;
+    }
+
+    return true;
+  } catch (error) {
+    console.error('Error updating workout date:', error);
+    return false;
+  }
+};
+
+/**
+ * Track workout set
  */
 export const trackWorkoutSet = async (
   workoutId: string,
@@ -18,13 +62,18 @@ export const trackWorkoutSet = async (
   }
 ) => {
   try {
-    // Implementation for tracking workout sets
-    console.log('Tracking workout set:', { workoutId, exerciseId, setData });
-    
-    const { data, error } = await supabase
+    const { data: userData } = await supabase.auth.getUser();
+    const userId = userData.user?.id;
+
+    if (!userId) {
+      console.error('No authenticated user found');
+      return false;
+    }
+
+    const { error } = await supabase
       .from('workout_set_completions')
       .insert({
-        user_id: (await supabase.auth.getUser()).data.user?.id,
+        user_id: userId,
         workout_completion_id: workoutId,
         workout_exercise_id: exerciseId,
         set_number: setData.set_number,
@@ -34,14 +83,13 @@ export const trackWorkoutSet = async (
         distance: setData.distance,
         duration: setData.duration,
         location: setData.location
-      })
-      .select();
-    
+      });
+
     if (error) {
       console.error('Error tracking workout set:', error);
       return false;
     }
-    
+
     return true;
   } catch (error) {
     console.error('Error tracking workout set:', error);
@@ -59,10 +107,7 @@ export const completeWorkout = async (
   rating?: number
 ) => {
   try {
-    // Implementation for completing a workout
-    console.log('Completing workout:', { workoutId, userId, notes, rating });
-    
-    const { data, error } = await supabase
+    const { error } = await supabase
       .from('workout_completions')
       .update({
         completed_at: new Date().toISOString(),
@@ -70,51 +115,16 @@ export const completeWorkout = async (
         rating: rating
       })
       .eq('id', workoutId)
-      .eq('user_id', userId)
-      .select();
-    
+      .eq('user_id', userId);
+
     if (error) {
       console.error('Error completing workout:', error);
       return false;
     }
-    
+
     return true;
   } catch (error) {
     console.error('Error completing workout:', error);
-    return false;
-  }
-};
-
-/**
- * Save workout journal notes
- */
-export const saveWorkoutJournalNotes = async (
-  userId: string,
-  notes: string,
-  emojiPrefix?: string
-) => {
-  try {
-    // Implementation for saving workout journal notes
-    console.log('Saving workout journal notes:', { userId, notes, emojiPrefix });
-    
-    const { data, error } = await supabase
-      .from('client_notes')
-      .insert({
-        user_id: userId,
-        content: notes,
-        emoji: emojiPrefix,
-        entry_date: new Date().toISOString()
-      })
-      .select();
-    
-    if (error) {
-      console.error('Error saving workout journal notes:', error);
-      return false;
-    }
-    
-    return true;
-  } catch (error) {
-    console.error('Error saving workout journal notes:', error);
     return false;
   }
 };
